[android] refactor code into ResourceDesignerGenerator
XamarinAndroidBuild class is quite large, trying to split things up
